Refactor naming in coreblocks frontend. (#570)
