+++ test/structs_common/test_csr.py
@@ -8,7 +8,7 @@ from coreblocks.params.configurations import test_core_config
 from coreblocks.params.layouts import ExceptionRegisterLayouts
 from coreblocks.params.keys import AsyncInterruptInsertSignalKey, ExceptionReportKey
 from transactron.utils.dependencies import DependencyManager
-from coreblocks.frontend.decoder import OpType
+from coreblocks.params.optypes import OpType
 
 from ..common import *
